hashsum and its aliases
